패키지 분리 원칙 예시 추가함
