+++ app/src/main/AndroidManifest.xml
@@ -15,16 +15,16 @@
         android:theme="@style/Theme.MyApplication"
         tools:targetApi="31">
         <activity
-            android:name=".ActivityJoin"
+            android:name=".ui.fragment.ActivityJoin"
             android:exported="false" />
         <activity
             android:name=".ActivityMain"
             android:exported="false" />
         <activity
-            android:name=".ActivityLogin"
+            android:name=".ui.fragment.ActivityLogin"
             android:exported="false" />
         <activity
-            android:name=".ActivitySplash"
+            android:name=".ui.fragment.ActivitySplash"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
