+++ app/src/main/java/com/cctv/heygongc/ui/analysis/AnalysisBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.cctv.heygongc.fragment
+package com.cctv.heygongc.ui.analysis
 
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
