+++ app/src/main/java/com/cctv/heygongc/ui/analysis/AnalysisFragment.kt
@@ -1,21 +1,16 @@
-package com.cctv.heygongc.fragment
+package com.cctv.heygongc.ui.analysis
 
-import android.annotation.SuppressLint
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearSmoothScroller
 import androidx.recyclerview.widget.RecyclerView
-import com.cctv.heygongc.R
-import com.cctv.heygongc.adapter.DateRecyclerViewAdapter
+import com.cctv.heygongc.ui.adapter.DateRecyclerViewAdapter
 import com.cctv.heygongc.data.TopDateData
 import com.cctv.heygongc.databinding.FragmentAnalysisBinding
-import com.cctv.heygongc.databinding.FragmentMonitoringBinding
 import java.util.Calendar
-import java.util.Date
 
 class AnalysisFragment : Fragment() {
     private var mBinding: FragmentAnalysisBinding? = null
