+++ app/src/main/java/com/cctv/heygongc/ui/fragment/ActivityJoin.kt
@@ -1,4 +1,4 @@
-package com.cctv.heygongc
+package com.cctv.heygongc.ui.fragment
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
