+++ app/src/main/java/com/cctv/heygongc/ui/fragment/ActivitySplash.kt
@@ -1,12 +1,12 @@
-package com.cctv.heygongc
+package com.cctv.heygongc.ui.fragment
 
-import android.content.Context
 import android.content.Intent
 import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.WindowManager
 import androidx.core.view.WindowCompat
+import com.cctv.heygongc.R
 import java.util.Timer
 import java.util.TimerTask
