+++ app/src/main/java/com/cctv/heygongc/ui/fragment/MonitoringFragment.kt
@@ -1,4 +1,4 @@
-package com.cctv.heygongc.fragment
+package com.cctv.heygongc.ui.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,12 +6,11 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.cctv.heygongc.R
-import com.cctv.heygongc.databinding.FragmentAnalysisBinding
 import com.cctv.heygongc.databinding.FragmentMonitoringBinding
-import com.cctv.heygongc.databinding.FragmentPremiumBinding
 
-class PremiumFragment : Fragment() {
-    private var mBinding: FragmentPremiumBinding? = null
+class MonitoringFragment : Fragment() {
+
+    private var mBinding: FragmentMonitoringBinding? = null
     private val binding get() = mBinding!!
 
 
@@ -19,12 +18,13 @@ class PremiumFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        mBinding = FragmentPremiumBinding.inflate(inflater, container, false)
+        mBinding = FragmentMonitoringBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onDestroyView() {
         mBinding = null
         super.onDestroyView()
+
     }
 }
