+++ app/src/main/java/com/cctv/heygongc/ui/fragment/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.cctv.heygongc.fragment
+package com.cctv.heygongc.ui.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,11 +6,11 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.cctv.heygongc.R
-import com.cctv.heygongc.databinding.FragmentMonitoringBinding
+import com.cctv.heygongc.databinding.FragmentPremiumBinding
+import com.cctv.heygongc.databinding.FragmentProfileBinding
 
-class MonitoringFragment : Fragment() {
-
-    private var mBinding: FragmentMonitoringBinding? = null
+class ProfileFragment : Fragment() {
+    private var mBinding: FragmentProfileBinding? = null
     private val binding get() = mBinding!!
 
 
@@ -18,13 +18,13 @@ class MonitoringFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        mBinding = FragmentMonitoringBinding.inflate(inflater, container, false)
+
+        mBinding = FragmentProfileBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onDestroyView() {
         mBinding = null
         super.onDestroyView()
-
     }
 }
