+++ app/src/main/res/layout/activity_join.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActivityJoin"
+    tools:context=".ui.fragment.ActivityJoin"
     android:orientation="vertical"
     android:padding="20dp">
