+++ app/src/main/res/layout/fragment_analysis.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragment.MonitoringFragment"
+    tools:context=".ui.fragment.MonitoringFragment"
     android:padding="15dp">
 
     <TextView
